+++ apps/web/client/src/app/project/[id]/_hooks/use-ai-sdk.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { useEditorEngine } from "@/components/store";
+import { useChat } from '@ai-sdk/react';
+
+export function useAiSdk() {
+    const editorEngine = useEditorEngine();
+    const { messages, input, handleInputChange, handleSubmit } = useChat();
+
+
+}
