+++ packages/utility/src/index.ts
@@ -4,6 +4,7 @@ export * from './color';
 export * from './errors';
 export * from './font';
 export * from './id';
+export * from './image';
 export * from './math';
 export * from './null';
 export * from './string';
@@ -11,4 +12,3 @@ export * from './tailwind';
 export * from './time';
 export * from './unit';
 export * from './urls';
-
